test(activerecord): merge the colliding postgresql/schema-statements-class test-file pair
Both files are trails inventions: rubyToConventionTs only maps
cases/connection_adapters/*_test.rb under connection-adapters/, so no Rails
port ever lands at either path. Fold the plain file into its .trails.test.ts
sibling, dropping the duplicated withSchemaStatements helper, and repoint the
require-canonical-rebuild exclude.

The other three colliding pairs are filed as follow-up stories; each costs
more than the LOC ceiling on its own.
