+++ eslint/require-canonical-rebuild-exclude.json
@@ -16,6 +16,6 @@
     "packages/activerecord/src/transactions.trails.test.ts"
   ],
   "nonExecuting": [
-    "packages/activerecord/src/connection-adapters/postgresql/schema-statements-class.test.ts"
+    "packages/activerecord/src/connection-adapters/postgresql/schema-statements-class.trails.test.ts"
   ]
 }
